requirements.txt
# ============================================================
# Quality Vision — Requirements
# ============================================================
# Instale com: pip install -r requirements.txt
# Para GPU (RTX 3050): depois execute:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# ============================================================

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Data Processing
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
opencv-python>=4.8.0

# Utilities
tqdm>=4.66.0
pydantic>=2.0.0
aiofiles>=23.0.0

# Database (logs)
# SQLite vem embutido no Python — sem instalação necessária

# Visualization (opcional, para notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Notebooks (opcional)
jupyter>=1.0.0
ipykernel>=6.0.0

kaggle>=1.6.0